feat: `cli` 支持修改 `package.json` 中的 `description` 字段
